requirements.txt
selenium==4.11.2
webdriver-manager==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
SQLite-utils==3.35
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
psutil==5.9.5
pywin32==306; platform_system=="Windows"